Renamed files and file references
